requirements.txt
requests
appdirs
install-jdk
click
pyaxmlparser
pyelftools
colorama